master refactor infrastructure and clean in pom.xml

+++ src/main/java/com/demo/spring/errorhandler/domain/order/port/OrderRepository.java
@@ -2,9 +2,8 @@ package com.demo.spring.errorhandler.domain.order.port;
 
 import com.demo.spring.errorhandler.domain.order.model.Order;
 
-import java.util.List;
 import java.util.Optional;
 
 public interface OrderRepository {
-    Optional<Order> findById(Long orderId);
+    Optional<Order> findById(Integer orderId);
 }
